refactor(back-top): sfc to tsx
