+++ src/back-top/back-top.en-US.md
@@ -1,15 +1,16 @@
 :: BASE_DOC ::
 
 ## API
+
 ### BackTop Props
 
 name | type | default | description | required
 -- | -- | -- | -- | --
 fixed | Boolean | true | \- | N
-icon | Slot / Function / Boolean | true | Typescript：`TNode`。[see more ts definition](https://github.com/Tencent/tdesign-mobile-vue/blob/develop/src/common.ts) | N
+icon | Boolean / Slot / Function | true | Typescript：`boolean \| TNode`。[see more ts definition](https://github.com/Tencent/tdesign-mobile-vue/blob/develop/src/common.ts) | N
 target | Function | - | Typescript：`() => HTMLElement` | N
 text | String | '' | \- | N
-theme | String | round | options：round/half-round/round-dark/half-round-dark | N
+theme | String | round | options: round/half-round/round-dark/half-round-dark | N
 onToTop | Function |  | Typescript：`() => void`<br/> | N
 
 ### BackTop Events
@@ -18,8 +19,8 @@ name | params | description
 -- | -- | --
 to-top | \- | \-
 
+### CSS 变量
 
-### CSS Variables
 The component provides the following CSS variables, which can be used to customize styles.
 Name | Default Value | Description 
 -- | -- | --
@@ -30,4 +31,4 @@ Name | Default Value | Description
 --td-back-top-round-border-radius | @radius-circle | - 
 --td-back-top-round-color | @font-gray-1 | - 
 --td-back-top-round-dark-bg-color | @gray-color-14 | - 
---td-back-top-round-dark-color | @font-white-1 | - 
+--td-back-top-round-dark-color | @font-white-1 | -
